requirements.txt
PyQt6
pyautogui
requests
anthropic
python-dotenv
pillow
numpy
qtawesome
SpeechRecognition
pyttsx3
keyboard
pyaudio
google-genai
